requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
websocket-client==1.6.4
python-dotenv==1.0.0
supabase==2.3.4
gotrue<2.0,>=1.3
openai==1.12.0
requests==2.31.0
python-socketio==5.10.0
flask-socketio==5.3.6
pydub==0.25.1

